+++ frontends/mit-open/src/pages/HomePage/HomePage.tsx
@@ -1,14 +1,28 @@
 import React from "react"
 import { Container, styled } from "ol-components"
 import HeroSearch from "./HeroSearch"
-import FeaturedResourcesSection from "./FeaturedResourcesSection"
-import MediaSection from "./MediaSection"
 import BrowseTopicsSection from "./BrowseTopicsSection"
 import NewsEventsSection from "./NewsEventsSection"
+import ResourceCarousel from "@/page-components/ResourceCarousel/ResourceCarousel"
+import * as carousels from "./carousels"
 
 const FullWidthBackground = styled.div({
   background: "linear-gradient(0deg, #FFF 0%, #E7EBEE 100%);",
 })
+
+const FeaturedCoursesCarousel = styled(ResourceCarousel)(({ theme }) => ({
+  margin: "80px 0",
+  [theme.breakpoints.down("sm")]: {
+    marginTop: "0px",
+    marginBottom: "32px",
+  },
+}))
+const MediaCarousel = styled(ResourceCarousel)(({ theme }) => ({
+  margin: "80px 0",
+  [theme.breakpoints.down("sm")]: {
+    margin: "40px 0",
+  },
+}))
 
 const HomePage: React.FC = () => {
   return (
@@ -16,11 +30,14 @@ const HomePage: React.FC = () => {
       <FullWidthBackground>
         <Container>
           <HeroSearch />
-          <FeaturedResourcesSection />
+          <FeaturedCoursesCarousel
+            title="Featured Courses"
+            config={carousels.FEATURED_RESOURCES_CAROUSEL}
+          />
         </Container>
       </FullWidthBackground>
-      <Container sx={{ marginTop: "24px" }}>
-        <MediaSection />
+      <Container>
+        <MediaCarousel title="Media" config={carousels.MEDIA_CAROUSEL} />
       </Container>
       <BrowseTopicsSection />
       <NewsEventsSection />
